# improve `interaction_matrix_exp!` performance