Using the first commit: (w/ TimerOutputs.jl)
```julia
julia> model = HubbardModelAttractive(L=5, t=1.0, U=0.8, dims=2)
julia> dqmc = DQMC(model, beta=1.0)
julia> run!(dqmc, thermalization=1000, sweeps=1000)
julia> MonteCarlo.to
 ────────────────────────────────────────────────────────────────────────────────────────
                                                 Time                   Allocations      
                                         ──────────────────────   ───────────────────────
            Tot / % measured:                 38.5s / 15.5%           11.8GiB / 74.4%    

 Section                         ncalls     time   %tot     avg     alloc   %tot      avg
 ────────────────────────────────────────────────────────────────────────────────────────
 run!                                 1    5.99s   100%   5.99s   8.78GiB  100%   8.78GiB
   interaction_matrix_exp! old     116k    617ms  10.3%  5.32μs    419MiB  4.66%  3.70KiB
   interaction_matrix_exp! new     116k   33.2ms  0.55%   286ns     0.00B  0.00%    0.00B
 ────────────────────────────────────────────────────────────────────────────────────────
```

The calculation of `lambda` is about 25ns on my machine, so not very relevant for larger systems. Another improvement can be made by only overwriting the diagonal entries, but I'm not sure if that's save to do. Using a `Diagonal` matrix as the target is also slightly faster. 

See ![here](https://gist.github.com/ffreyer/5322772e09456399500de61ed45a0c04) for some independent benchmarks.